Show the Goto menu as fallback when no file is found.
Sublime text doesn't allow us to query the file database of the project.
Show the Goto menu with the common base between extensions to give at least
that to the user instead of letting him do it manually if the project keeps
headers in a separate directory from the sources.

This should help for issue #1.